Separando en un nuevo controlador lo referente a TeamsGroup
