+++ frontend/server/nginx.rewrites
@@ -56,9 +56,7 @@ rewrite ^/course/([a-zA-Z0-9_+-]+)/students/?$ /course/students.php?course=$1 la
 rewrite ^/course/list/(student|public)/?$ /course/singlelist.php?course_type=$1 last;
 rewrite ^/group/?$ /group/list.php last;
 rewrite ^/group/new/?$ /group/new.php last;
-rewrite ^/group/new_teams_group/?$ /group/new.php?for_teams=true last;
 rewrite ^/group/([a-zA-Z0-9_+:-]+)/edit/?$ /group/edit.php?group=$1 last;
-rewrite ^/group/([a-zA-Z0-9_+:-]+)/edit_teams_group/?$ /group/edit_teams_group.php?team_group_alias=$1 last;
 rewrite ^/group/([a-zA-Z0-9_+:-]+)/scoreboard/([a-zA-Z0-9_+-]+)/?$ /groupscoreboard.php?group=$1&scoreboard=$2 last;
 rewrite ^/group/([a-zA-Z0-9_+:-]+)/scoreboard/([a-zA-Z0-9_+-]+)/edit/?$ /group/scoreboardedit.php?group=$1&scoreboard=$2 last;
 rewrite ^/group/([a-zA-Z0-9_+:-]+)/stats/?$ /groupstats.php?group=$1 last;
@@ -101,6 +99,8 @@ rewrite ^/schools/profile/([0-9]+)/?$ /schools/profile.php?school_id=$1 last;
 rewrite ^/scoreboardmerge/?$ /contests/scoreboardmerge.php last;
 rewrite ^/submissions/?$ /submissions/list.php last;
 rewrite ^/submissions/([a-zA-Z0-9_+.-]+)/?$ /submissions/user_list.php?username=$1 last;
+rewrite ^/teamsgroup/new/?$ /teamsgroup/new.php last;
+rewrite ^/teamsgroup/([a-zA-Z0-9_+:-]+)/edit/?$ /teamsgroup/edit.php?teams_group_alias=$1 last;
 rewrite ^/user/emailedit/([a-zA-Z0-9_]+)/?$ /users/emailedit.php?username=$1 last;
 rewrite ^/user/verifyemail/([a-zA-Z0-9_]+)/redirecttointerview/([a-zA-Z0-9_+-]+)/?$ /userverifyemail.php?id=$1&redirecttointerview=$2 last;
 rewrite ^/user/verifyemail/([a-zA-Z0-9_]+)/?$ /userverifyemail.php?id=$1 last;
