+++ frontend/server/src/DAO/TeamGroups.php
@@ -46,4 +46,29 @@ class TeamGroups extends \OmegaUp\DAO\Base\TeamGroups {
         }
         return new \OmegaUp\DAO\VO\TeamGroups($rs);
     }
+
+    /**
+     * Returns all teams groups that a user can manage.
+     * @param int $userId
+     * @param int $identityId
+     * @return list<array{alias: string, create_time: \OmegaUp\Timestamp, description: null|string, name: string}>
+     */
+    final public static function getAllTeamsGroupsAdminedByUser(int $userId) {
+        $sql = 'SELECT
+                    DISTINCT tg.alias,
+                    tg.create_time,
+                    tg.description,
+                    tg.name
+                FROM
+                    `Team_Groups` AS tg
+                INNER JOIN
+                    ACLs AS a ON a.acl_id = tg.acl_id
+                WHERE
+                    a.owner_id = ?
+                ORDER BY
+                    tg.create_time DESC;';
+
+        /** @var list<array{alias: string, create_time: \OmegaUp\Timestamp, description: null|string, name: string}> */
+        return \OmegaUp\MySQLConnection::getInstance()->GetAll($sql, [$userId]);
+    }
 }
