+++ frontend/templates/en.lang
@@ -660,8 +660,8 @@ omegaupTitleSchoolsRank = "School ranking"
 omegaupTitleScoreboardmerge = "Scoreboardmerge"
 omegaupTitleStudentsProgress = "Students' progress"
 omegaupTitleSupportDashboard = "Support dashboard"
-omegaupTitleTeamGroupsEdit = "Edit team group"
-omegaupTitleTeamGroupsNew = "New teams group"
+omegaupTitleTeamsGroupEdit = "Edit teams group"
+omegaupTitleTeamsGroupNew = "New teams group"
 omegaupTitleUpdatePrivileges = "Update privileges"
 omegaupTitleUsersRank = "User ranking"
 onlyDemotionsSupported = "Resolution only supported for demotions for now."
