+++ frontend/templates/es.lang
@@ -660,8 +660,8 @@ omegaupTitleSchoolsRank = "Ranking de escuelas"
 omegaupTitleScoreboardmerge = "Scoreboard Merge"
 omegaupTitleStudentsProgress = "Progreso de los estudiantes"
 omegaupTitleSupportDashboard = "Tablero de soporte"
-omegaupTitleTeamGroupsEdit = "Editar grupo de equipos"
-omegaupTitleTeamGroupsNew = "Nuevo grupo de equipos"
+omegaupTitleTeamsGroupEdit = "Editar grupo de equipos"
+omegaupTitleTeamsGroupNew = "Nuevo grupo de equipos"
 omegaupTitleUpdatePrivileges = "Actualizar privilegios"
 omegaupTitleUsersRank = "Ranking de usuarios"
 onlyDemotionsSupported = "Solamente se pueden resolver demociones por ahora."
