+++ frontend/templates/pt.lang
@@ -660,8 +660,8 @@ omegaupTitleSchoolsRank = "Ranking de escolas"
 omegaupTitleScoreboardmerge = "União de paineis de avaliação"
 omegaupTitleStudentsProgress = "Progresso dos alunos"
 omegaupTitleSupportDashboard = "Painel de Suporte"
-omegaupTitleTeamGroupsEdit = "Editar grupo de equipes"
-omegaupTitleTeamGroupsNew = "Novo grupo de equipes"
+omegaupTitleTeamsGroupEdit = "Editar grupo de equipes"
+omegaupTitleTeamsGroupNew = "Novo grupo de equipes"
 omegaupTitleUpdatePrivileges = "Atualizar privilégios"
 omegaupTitleUsersRank = "Ranking de usuários"
 onlyDemotionsSupported = "Resolução suportada apenas para demotions para agora."
