+++ frontend/www/group/new.php
@@ -3,7 +3,11 @@ namespace OmegaUp;
 require_once(dirname(__DIR__, 2) . '/server/bootstrap.php');
 
 \OmegaUp\UITools::render(
-    fn (\OmegaUp\Request $r) => \OmegaUp\Controllers\Group::getGroupNewDetailsForTypeScript(
-        $r,
-    )
+    fn (\OmegaUp\Request $r) => [
+        'smartyProperties' => [
+            'payload' => [],
+            'title' => new \OmegaUp\TranslationString('omegaupTitleGroupsNew'),
+        ],
+        'entrypoint' => 'group_new',
+    ]
 );
