+++ frontend/www/js/omegaup/api.ts
@@ -1081,17 +1081,9 @@ export const Group = {
     messages.GroupCreateScoreboardRequest,
     messages.GroupCreateScoreboardResponse
   >('/api/group/createScoreboard/'),
-  createTeamGroup: apiCall<
-    messages.GroupCreateTeamGroupRequest,
-    messages.GroupCreateTeamGroupResponse
-  >('/api/group/createTeamGroup/'),
   details: apiCall<messages.GroupDetailsRequest, messages.GroupDetailsResponse>(
     '/api/group/details/',
   ),
-  detailsForTeams: apiCall<
-    messages.GroupDetailsForTeamsRequest,
-    messages.GroupDetailsForTeamsResponse
-  >('/api/group/detailsForTeams/'),
   list: apiCall<messages.GroupListRequest, messages.GroupListResponse>(
     '/api/group/list/',
   ),
@@ -1717,6 +1709,17 @@ export const Tag = {
   list: apiCall<messages.TagListRequest, messages.TagListResponse>(
     '/api/tag/list/',
   ),
+};
+
+export const TeamsGroup = {
+  create: apiCall<
+    messages.TeamsGroupCreateRequest,
+    messages.TeamsGroupCreateResponse
+  >('/api/teamsGroup/create/'),
+  details: apiCall<
+    messages.TeamsGroupDetailsRequest,
+    messages.TeamsGroupDetailsResponse
+  >('/api/teamsGroup/details/'),
 };
 
 export const Time = {
