+++ frontend/www/js/omegaup/lang.en.json
@@ -661,8 +661,8 @@
 	"omegaupTitleScoreboardmerge": "Scoreboardmerge",
 	"omegaupTitleStudentsProgress": "Students' progress",
 	"omegaupTitleSupportDashboard": "Support dashboard",
-	"omegaupTitleTeamGroupsEdit": "Edit team group",
-	"omegaupTitleTeamGroupsNew": "New teams group",
+	"omegaupTitleTeamsGroupEdit": "Edit teams group",
+	"omegaupTitleTeamsGroupNew": "New teams group",
 	"omegaupTitleUpdatePrivileges": "Update privileges",
 	"omegaupTitleUsersRank": "User ranking",
 	"onlyDemotionsSupported": "Resolution only supported for demotions for now.",
