+++ frontend/www/js/omegaup/lang.es.ts
@@ -662,8 +662,8 @@ const translations: { [key: string]: string; } = {
   omegaupTitleScoreboardmerge: "Scoreboard Merge",
   omegaupTitleStudentsProgress: "Progreso de los estudiantes",
   omegaupTitleSupportDashboard: "Tablero de soporte",
-  omegaupTitleTeamGroupsEdit: "Editar grupo de equipos",
-  omegaupTitleTeamGroupsNew: "Nuevo grupo de equipos",
+  omegaupTitleTeamsGroupEdit: "Editar grupo de equipos",
+  omegaupTitleTeamsGroupNew: "Nuevo grupo de equipos",
   omegaupTitleUpdatePrivileges: "Actualizar privilegios",
   omegaupTitleUsersRank: "Ranking de usuarios",
   onlyDemotionsSupported: "Solamente se pueden resolver demociones por ahora.",
