+++ frontend/www/js/omegaup/lang.pt.json
@@ -661,8 +661,8 @@
 	"omegaupTitleScoreboardmerge": "Uni\u00e3o de paineis de avalia\u00e7\u00e3o",
 	"omegaupTitleStudentsProgress": "Progresso dos alunos",
 	"omegaupTitleSupportDashboard": "Painel de Suporte",
-	"omegaupTitleTeamGroupsEdit": "Editar grupo de equipes",
-	"omegaupTitleTeamGroupsNew": "Novo grupo de equipes",
+	"omegaupTitleTeamsGroupEdit": "Editar grupo de equipes",
+	"omegaupTitleTeamsGroupNew": "Novo grupo de equipes",
 	"omegaupTitleUpdatePrivileges": "Atualizar privil\u00e9gios",
 	"omegaupTitleUsersRank": "Ranking de usu\u00e1rios",
 	"onlyDemotionsSupported": "Resolu\u00e7\u00e3o suportada apenas para demotions para agora.",
