+++ frontend/www/teamsgroup/edit.php
@@ -3,7 +3,7 @@ namespace OmegaUp;
 require_once(dirname(__DIR__, 2) . '/server/bootstrap.php');
 
 \OmegaUp\UITools::render(
-    fn (\OmegaUp\Request $r) => \OmegaUp\Controllers\Group::getTeamGroupEditDetailsForTypeScript(
+    fn (\OmegaUp\Request $r) => \OmegaUp\Controllers\TeamsGroup::getTeamGroupEditDetailsForTypeScript(
         $r
     )
 );
