+++ frontend/www/teamsgroup/new.php
@@ -6,8 +6,10 @@ require_once(dirname(__DIR__, 2) . '/server/bootstrap.php');
     fn (\OmegaUp\Request $r) => [
         'smartyProperties' => [
             'payload' => [],
-            'title' => new \OmegaUp\TranslationString('omegaupTitleGroupsNew'),
+            'title' => new \OmegaUp\TranslationString(
+                'omegaupTitleTeamsGroupNew'
+            ),
         ],
-        'entrypoint' => 'group_new',
+        'entrypoint' => 'teams_group_new',
     ]
 );
